Makefile: Add `-gnatw_R` to suppress spurious warning
Looks like there's a new warning option that complains when the order of
component declarations, component clauses, and bit-level layout do not
all agree. However, we also enable `warnings are errors`. Plus, we use
mismatched component representation clauses on purpose in libgfxinit's
epilepsy-inducing gfxtest program to have RGB ordering on hardware that
works with BGR ordering. Hardware is cursed, so silence this warning.

Change-Id: I1565e3e834cd47b7f53fd2202b38fdf01a0b7619
Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://review.coreboot.org/c/libhwbase/+/43556
Tested-by: Nico Huber <nico.h@gmx.de>
Reviewed-by: Nico Huber <nico.h@gmx.de>