hw-debug: Place global variables in the .bss section
It places the Middle_Of_Line and Register_Write_Delay_Nanoseconds
variables in the .bss section. As a result they are set at runtime to
their desired default value which is respectively False and 0.

It helps to integrate the libhwbase library debug package in
environments where global initialized variables are not
supported such as coreboot romstage.

This change does not impact GNATprove as this package has SPARK_Mode
disabled.

BUG=b:252792591
BRANCH=firmware-brya-14505.B
TEST=Working as expected on skolas

Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Change-Id: Ib23699f3f3446219c34c60ccd9987346e17769f2
Reviewed-on: https://review.coreboot.org/c/libhwbase/+/69854
Tested-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>